requirements.txt
Pillow>=10.0.0  # For image processing
imagehash>=4.3.1  # For image comparison
tqdm>=4.65.0     # For progress bars
pillow-heif>=0.15.0  # For HEIC support 
piexif>=1.1.3    # For EXIF metadata handling 